test: place c06 CLI process witnesses in integration lane
