--- tests/functional/transport/cli/process/fixtures_test.go
@@ -1,79 +1,0 @@
-package process_test
-
-import (
-	"context"
-	"encoding/json"
-	"errors"
-	"os"
-	"path/filepath"
-	"testing"
-
-	"github.com/portpowered/infinite-you/internal/builtcliacceptance"
-	"github.com/portpowered/infinite-you/pkg/services/workers"
-)
-
-type operatorConfigOutcome struct {
-	ConfigPath string
-}
-
-func initializeOperatorConfig(
-	t testing.TB,
-	ctx context.Context,
-	binaryPath string,
-	session *builtcliacceptance.Session,
-	scenario string,
-) operatorConfigOutcome {
-	t.Helper()
-
-	configPath := filepath.Join(session.HomeDir, ".you-agent-factory", "config.json")
-	missingFactory := filepath.Join(session.WorkDir, "missing-initialization-factory.json")
-	result, err := runBuiltYouBinary(ctx, binaryPath, session, "run", "--factory", missingFactory)
-	if err == nil {
-		t.Fatalf("%s: run missing Factory unexpectedly succeeded: stdout=%q stderr=%q", scenario, result.Stdout, result.Stderr)
-	}
-	if _, err := os.Stat(configPath); errors.Is(err, os.ErrNotExist) {
-		t.Fatalf("%s: initializer-owned config missing at %s: %v", scenario, configPath, err)
-	}
-	return operatorConfigOutcome{ConfigPath: configPath}
-}
-
-func writeAcceptingGoalMockWorkers(t *testing.T) string {
-	t.Helper()
-
-	data, err := json.Marshal(workers.MockWorkersConfig{
-		UnmatchedDispatchPolicy: workers.MockWorkerUnmatchedDispatchPolicyPassthrough,
-		MockWorkers: []workers.MockWorkerConfig{
-			{WorkerName: "goal-planner", WorkstationName: "plan-goal", RunType: workers.MockWorkerRunTypeAccept},
-			{WorkerName: "goal-executor", WorkstationName: "execute-goal", RunType: workers.MockWorkerRunTypeAccept},
-			{WorkerName: "goal-checker", WorkstationName: "check-goal", RunType: workers.MockWorkerRunTypeAccept},
-			{WorkerName: "goal-reviewer", WorkstationName: "review-goal", RunType: workers.MockWorkerRunTypeAccept},
-		},
-	})
-	if err != nil {
-		t.Fatalf("marshal accepting mock workers: %v", err)
-	}
-	path := filepath.Join(t.TempDir(), "accepting-mock-workers.json")
-	if err := os.WriteFile(path, data, 0o600); err != nil {
-		t.Fatalf("write accepting mock workers: %v", err)
-	}
-	return path
-}
-
-func writeRejectingGoalMockWorkers(t *testing.T) string {
-	t.Helper()
-
-	data, err := json.Marshal(workers.MockWorkersConfig{MockWorkers: []workers.MockWorkerConfig{
-		{WorkerName: "goal-planner", WorkstationName: "plan-goal", RunType: workers.MockWorkerRunTypeReject},
-		{WorkerName: "goal-executor", WorkstationName: "execute-goal", RunType: workers.MockWorkerRunTypeReject},
-		{WorkerName: "goal-checker", WorkstationName: "check-goal", RunType: workers.MockWorkerRunTypeReject},
-		{WorkerName: "goal-reviewer", WorkstationName: "review-goal", RunType: workers.MockWorkerRunTypeReject},
-	}})
-	if err != nil {
-		t.Fatalf("marshal rejecting mock workers: %v", err)
-	}
-	path := filepath.Join(t.TempDir(), "rejecting-mock-workers.json")
-	if err := os.WriteFile(path, data, 0o600); err != nil {
-		t.Fatalf("write rejecting mock workers: %v", err)
-	}
-	return path
-}
